Conflicts:
	PotreeConverter/include/FlatBufferReader.hpp
	PotreeConverter/src/FlatBufferReader.cpp

+++ PotreeConverter/include/BINPointReader.hpp
@@ -43,7 +43,7 @@ public:
 
 	AABB getAABB();
 
-	long long numPoints();
+	int64_t numPoints();
 
 	void close();
