+++ PotreeConverter/include/LASPointReader.h
@@ -153,7 +153,7 @@ public:
 
 	AABB getAABB();
 
-	long long numPoints();
+	int64_t numPoints();
 
 	void close();
